Allow integration tests against k8s cluster
Extends test support in `monolithic_integration_test` so that tests can
be run against Janus instances already running in a Kubernetes cluster.
The cluster could either be running locally in Kind, or remotely in a
cloud platform.

It's assumed that two Janus instances are running in the cluster, in
distinct Kubernetes namespaces. It's also assumed that there exist
Kubernetes services named `aggregator` and `postgresql` for the DAP API
and the datastore, respectively.

Parameters needed to connect to the Kubernetes cluster are provided to
the test in environment variables, documented in `JanusPair::new`. If
the environment variables are empty, then the test is run against
ephemeral, in-process Janus components.

Part of https://github.com/divviup/janus-ops/issues/39,
https://github.com/divviup/janus-ops/issues/141